fix calculating the index from ReferenceType and vise versa, because it was not correctly implemented
