+++ src/ui/ReferenceView.cpp
@@ -15,8 +15,6 @@
 #include "dialogs/MergeDialog.h"
 #include "git/Branch.h"
 #include "git/Repository.h"
-#include "git/Signature.h"
-#include "git/Tag.h"
 #include "git/TagRef.h"
 #include "log/LogEntry.h"
 #include <QAbstractItemModel>
@@ -31,283 +29,11 @@
 #include <QTreeView>
 #include <QVBoxLayout>
 
+#include "ReferenceModel.h"
+
 namespace {
 
 const int kHeight = 200;
-const QString kNowrapFmt = "<span style='white-space: nowrap'>%1</span>";
-
-bool refComparator(const git::Reference &lhs, const git::Reference &rhs) {
-  git::Commit lhsCommit = lhs.target();
-  git::Commit rhsCommit = rhs.target();
-  return (lhsCommit.isValid() && rhsCommit.isValid() &&
-          lhsCommit.committer().date() > rhsCommit.committer().date());
-}
-
-enum ReferenceType { HEADER = 0, Branches = 1, Remotes = 2, Tags = 3 };
-
-class ReferenceModel : public QAbstractItemModel {
-  Q_OBJECT
-
-public:
-  struct ReferenceList {
-    QString name;
-    QList<git::Reference> refs;
-  };
-
-  ReferenceModel(const git::Repository &repo, ReferenceView::Kinds kinds,
-                 bool filterCurrentCommit = false, QObject *parent = nullptr)
-      : QAbstractItemModel(parent), mRepo(repo), mKinds(kinds),
-        mFilterCurrentCommit(filterCurrentCommit) {
-    git::RepositoryNotifier *notifier = repo.notifier();
-    connect(notifier, &git::RepositoryNotifier::referenceAdded, this,
-            &ReferenceModel::update);
-    connect(notifier, &git::RepositoryNotifier::referenceRemoved, this,
-            &ReferenceModel::update);
-    connect(notifier, &git::RepositoryNotifier::referenceUpdated, this,
-            &ReferenceModel::update);
-  }
-
-  static int referenceTypeToIndex(ReferenceType t) { return t - 1; }
-
-  static int indexToReferenceType(int index) { return index + 1; }
-
-  void setCommit(const git::Commit &commit) {
-    mCommit = commit;
-    update();
-  }
-
-  void update() {
-    beginResetModel();
-
-    mRefs.clear();
-
-    // Add detached head.
-    git::Reference detachedHead;
-    if (mKinds & ReferenceView::DetachedHead) {
-      git::Reference head = mRepo.head();
-      if (head.isValid() && !head.isBranch()) {
-        if (!mFilterCurrentCommit ||
-            head.annotatedCommit().commit() == mCommit) {
-          detachedHead = head;
-        }
-      }
-    }
-
-    // Add local branches.
-    if (mKinds & ReferenceView::LocalBranches) {
-      QList<git::Reference> branches;
-      foreach (const git::Branch &branch, mRepo.branches(GIT_BRANCH_LOCAL)) {
-        qDebug() << "ReferenceView: Local branches: " << branch.name();
-        const bool branchOnCommit =
-            !mCommit.isValid() || branch.annotatedCommit().commit() == mCommit;
-        if ((!(mKinds & ReferenceView::ExcludeHead) || !branch.isHead()) &&
-            (branchOnCommit || !mFilterCurrentCommit))
-          branches.append(branch);
-      }
-
-      std::sort(branches.begin(), branches.end(), refComparator);
-
-      // Add top references.
-      if (detachedHead.isValid())
-        branches.prepend(detachedHead);
-      if (mKinds & ReferenceView::InvalidRef)
-        branches.prepend(git::Reference());
-
-      // Add bottom references.
-      if (mKinds & ReferenceView::Stash) {
-        if (git::Reference stash = mRepo.stashRef())
-          branches.append(stash);
-      }
-
-      mRefs.append({tr("Branches"), branches}); // First element in mRefs
-    }
-
-    // Add remote branches.
-    if (mKinds & ReferenceView::RemoteBranches) {
-      QList<git::Reference> remotes;
-      foreach (const git::Branch &branch, mRepo.branches(GIT_BRANCH_REMOTE)) {
-        // Filter remote HEAD branches.
-        qDebug() << "ReferenceView: Remote branches: " << branch.name();
-        const bool remoteOnCommit =
-            !mCommit.isValid() || branch.annotatedCommit().commit() == mCommit;
-        if (!branch.name().endsWith("HEAD") &&
-            (remoteOnCommit || !mFilterCurrentCommit))
-          remotes.append(branch);
-      }
-
-      std::sort(remotes.begin(), remotes.end(), refComparator);
-      if (mKinds & ReferenceView::InvalidRef)
-        remotes.prepend(git::Reference());
-      mRefs.append({tr("Remotes"), remotes}); // Second element in mRefs
-    }
-
-    // Add tags.
-    if (mKinds & ReferenceView::Tags) {
-      QList<git::Reference> tags;
-      foreach (const git::TagRef &tag, mRepo.tags()) {
-        qDebug() << "ReferenceView: Tags: " << tag.name();
-        const bool tagOnCommit =
-            !mCommit.isValid() || tag.annotatedCommit().commit() == mCommit;
-        if (tagOnCommit || !mFilterCurrentCommit)
-          tags.append(tag);
-      }
-
-      std::sort(tags.begin(), tags.end(), refComparator);
-      mRefs.append({tr("Tags"), tags}); // Third element in mRefs
-    }
-
-    endResetModel();
-  }
-
-  QModelIndex firstRemote() {
-    // Return first remote if available, otherwise an invalid modelIndex
-    for (auto &ref : mRefs) {
-      if (ref.name == tr("Remotes") && ref.refs.count() > 0) {
-        if (mKinds & ReferenceView::InvalidRef) {
-          if (ref.refs.count() > 1) {
-            // use the first valid ref after the invalid ref
-            return createIndex(1, 0, ReferenceType::Remotes);
-          }
-        } else
-          return createIndex(0, 0, ReferenceType::Remotes);
-        break; // Remotes are already found so no need to continue searching
-      }
-    }
-    return QModelIndex();
-  }
-
-  QModelIndex firstBranch() {
-    // Return first branch if available, otherwise an invalid modelIndex
-    for (auto &ref : mRefs) {
-      if (ref.name == tr("Branches") && ref.refs.count() > 0) {
-        if (mKinds & ReferenceView::InvalidRef) {
-          if (ref.refs.count() >
-              1) // use the first valid ref after the invalid ref
-            return createIndex(1, 0, ReferenceType::Branches);
-        } else
-          return createIndex(0, 0, ReferenceType::Branches);
-        break; // Remotes are already found so no need to continue searching
-      }
-    }
-    return QModelIndex();
-  }
-
-  QModelIndex firstTag() {
-    // Return first tag if available, otherwise an invalid modelIndex
-
-    for (auto &ref : mRefs) {
-      if (ref.name == tr("Tags") && ref.refs.count() > 0)
-        return createIndex(0, 0, ReferenceType::Tags);
-    }
-    return QModelIndex();
-  }
-
-  QModelIndex index(int row, int column,
-                    const QModelIndex &parent = QModelIndex()) const override {
-    if (!hasIndex(row, column, parent))
-      return QModelIndex();
-
-    // headers do not have a parent
-    bool header = !parent.isValid();
-    return createIndex(row, column,
-                       header ? 0 : indexToReferenceType(parent.row()));
-  }
-
-  QModelIndex parent(const QModelIndex &index) const override {
-    if (!index.isValid())
-      return QModelIndex();
-
-    // The header sections (Branches, Remotes, Tags) (Elements of mRefs)
-    // do not have any parents, only the refs in each mRefs element
-    const quintptr id = index.internalId();
-    if (id == HEADER)
-      return QModelIndex();
-    const auto row = referenceTypeToIndex(static_cast<ReferenceType>(id));
-    return createIndex(row, 0);
-  }
-
-  int rowCount(const QModelIndex &parent = QModelIndex()) const override {
-    if (!parent.isValid())
-      return mRefs.size();
-
-    if (parent.internalId() != HEADER)
-      return 0; // refs it self do not have childs
-
-    return mRefs.at(parent.row()).refs.size();
-  }
-
-  int columnCount(const QModelIndex &parent = QModelIndex()) const override {
-    return 1;
-  }
-
-  QVariant data(const QModelIndex &index,
-                int role = Qt::DisplayRole) const override {
-    // kinds
-    int row = index.row();
-    quintptr id = index.internalId();
-    if (id == HEADER)
-      return (role == Qt::DisplayRole) ? mRefs.at(row).name : QVariant();
-
-    auto refType = static_cast<ReferenceType>(id);
-
-    // refs
-    git::Reference ref = mRefs.at(referenceTypeToIndex(refType)).refs.at(row);
-    switch (role) {
-      case Qt::DisplayRole:
-        return ref.isValid() ? ref.name() : QString();
-
-      case Qt::ToolTipRole: {
-        if (!ref.isValid() || !ref.isTag())
-          return QVariant();
-
-        git::Tag tag = git::TagRef(ref).tag();
-        if (!tag.isValid())
-          return QVariant();
-
-        QStringList lines;
-        if (git::Signature signature = tag.tagger()) {
-          QString name = QString("<b>%1</b>").arg(signature.name());
-          QString email = QString("&lt;%1&gt;").arg(signature.email());
-          lines.append(kNowrapFmt.arg(QString("%1 %2").arg(name, email)));
-
-          QString date = signature.date().toString(Qt::DefaultLocaleLongDate);
-          lines.append(kNowrapFmt.arg(date));
-        }
-
-        QString msg = tag.message();
-        if (!msg.isEmpty())
-          lines.append(QString("<p>%1</p>").arg(msg));
-
-        return lines.join('\n');
-      }
-
-      case Qt::FontRole: {
-        QFont font = static_cast<QWidget *>(QObject::parent())->font();
-        font.setBold(ref.isValid() && ref.isHead());
-        return font;
-      }
-
-      case Qt::UserRole:
-        return QVariant::fromValue(ref);
-
-      default:
-        return QVariant();
-    }
-  }
-
-  QVariant headerData(int section, Qt::Orientation orientation,
-                      int role = Qt::DisplayRole) const override {
-    return QVariant();
-  }
-
-private:
-  git::Repository mRepo;
-  ReferenceView::Kinds mKinds;
-  bool mFilterCurrentCommit; // Filter out all branches, remotes, tags which are
-                             // related to the commit mCommit
-  QList<ReferenceList> mRefs;
-  git::Commit mCommit;
-};
 
 class FilterProxyModel : public QSortFilterProxyModel {
 public:
